Added full ES6 support
This changes the JavaScript parser to use acorn.js, allowing support for
all ES6 features. The parser will now add entries for ES6 classes, methods
and shorthand function properties. Additionally, it indents functions, and
if a function does not have a name, it can derive a name from variable names
or object properties that the function is assigned to.